changed matrix dimentions to unsigned long long

+++ include/matrix.h
@@ -2,7 +2,7 @@
 #define PDDP_MATRIX_H
 
 struct Matrix{
-    int rows, cols;
+    unsigned long long rows, cols;
     double* matrix;
 };
 
@@ -11,9 +11,9 @@ typedef struct Matrix Matrix;
 Matrix calculateAverageVector(Matrix);
 Matrix calculateMean(Matrix, Matrix);
 Matrix calculateAtA(Matrix);
-double calculateAtAElement(Matrix, int, int);
+double calculateAtAElement(Matrix, unsigned long long, unsigned long long);
 Matrix matrixMultiply(Matrix, Matrix);
-Matrix ones(int , int);
+Matrix ones(unsigned long long , unsigned long long);
 double norm(Matrix);
 Matrix divideMatrixByScalar(Matrix, double);
 Matrix matrixSubtract(Matrix, Matrix);
